Added more
Added more CSS (Left aligned)
Fixed markdown editor overlap
Added confirmations for serious events

+++ src/components/DeleteFileButton.js
@@ -1,16 +1,18 @@
 import DeleteFile from '../hooks/DeleteFile';
 import CheckUser from '../hooks/CheckUser';
 
-const DeleteFileButton = ({directory, pageName}) => {
+const DeleteFileButton = ({ directory, pageName }) => {
 
     const deletePage = async () => {
-        try {
-            await DeleteFile(`${directory}/${pageName}.md`, localStorage.getItem('userToken'));
-            alert('Successfully deleted page!');
-            window.location.href = `#category/${directory}`;
-        }
-        catch (error) {
-            alert('Failed to delete page: ', error);
+        if (window.confirm("Are you sure you want to delete this page? This action cannot be undone.")) {
+            try {
+                await DeleteFile(`${directory}/${pageName}.md`, localStorage.getItem('userToken'));
+                alert('Successfully deleted page!');
+                window.location.href = `#category/${directory}`;
+            }
+            catch (error) {
+                alert('Failed to delete page: ', error);
+            }
         }
     };
